Dataset: Phenology of flora of mediterranean high-mountains meadows (Sierra Nevada)

This dataset contains information about the phenology of flora in Mediterranean high-mountains meadows in the Sierra Nevada region.

Usage

borreguiles

Format

A data frame with 11002 observations and 37 variables:

id: The ID of the observation.

modified: The modification date of the observation.

language: The language of the observation.

institutionCode, collectionCode: The institution and collection code.

datasetName: The name of the dataset.

basisOfRecord: The basis of the record.

occurrenceID: The occurrence ID.

catalogNumber: The catalog number.

recordedBy, recordedByID: The name and ID of the person who recorded the observation.

eventDate: The date of the observation.

year, month, day: The year, month and day of the observation.

continent: The continent where the observation was made.

country: The country where the observation was made.

countryCode: The country code.

stateProvince: The state or province where the observation was made.

county: The county where the observation was made.

locality: The locality of the observation.

minimumElevationInMeters, maximumElevationInMeters: The minimum and maximum elevation in meters.

decimalLatitude, decimalLongitude: The decimal latitude and longitude of the observation.

geodeticDatum: The geodetic datum.

coordinateUncertaintyInMeters: The coordinate uncertainty in meters.

scientificName: The scientific name of the plant species.

kingdom, phylum, class, order, family: The kingdom, phylum, class, order, family and genus of the taxa.

specificEpithet: The specific epithet of the plant species.

infraspecificEpithet: The infraspecific epithet of the plant species.

scientificNameAuthorship: The scientific name authorship.

Source

Zamora R, Pérez-Luque AJ (2020). Dataset of Phenology of flora of mediterranean high-mountains meadows (Sierra Nevada). Version 1.5. Sierra Nevada Global Change Observatory. Andalusian Environmental Center, University of Granada, Regional Government of Andalusia. Occurrence dataset doi: 10.15468/qhqzub

Pérez-Luque AJ, Sánchez-Rojas CP, Zamora R, Pérez-Pérez R, Bonet FJ (2015). Dataset of Phenology of Mediterranean high-mountain meadows flora (Sierra Nevada, Spain). PhytoKeys 46: 89-107. doi: 10.3897/phytokeys.46.9116