taxonomic_cov

It generates the taxonomic coverage of the dataset

Usage

taxonomic_cov(
  x,
  category = c("scientificName", "kingdom", "phylum", "class", "order", "family",
    "subfamily", "genus", "all"),
  code_unknown = NA
)

Arguments

x

A dataframe

category

a character vector with taxonomic ranks defined by DarwinCore standard. #' @param code_unknown a custom string that identified the unclassified records at taxa level (default to NA). If the dataset contains a custom string to identify the unclassified taxa, please specify it here, e.g. "unclass.", "unknown". See taxonomic_freq()

Value

A named list of tibbles summarizing the taxonomic coverage for each category.

Details

It generates the taxonomic coverage of a occurrence dataset. For each category, the record numbers and relative frequencies of each taxa are computed. The category argument could be one of: scientificName, kingdom, phylum, class, order, family, subfamily and genus. For get coverage in all categories type all.

Examples

# Example usage
data(borreguiles)
# Calculate taxonomic coverage for scientificName and genus
taxonomic_cov(borreguiles, category = c("scientificName", "genus"))
#> Joining with `by = join_by(scientificName, n)`
#> Joining with `by = join_by(genus, n)`
#> $scientificName
#> # A tibble: 94 × 4
#>    scientificName                                             n freq_all  freq
#>    <chr>                                                  <int>    <dbl> <dbl>
#>  1 Nardus stricta L.                                       1029     9.35  9.35
#>  2 Carex nigra (L.) Reichard                                734     6.67  6.67
#>  3 Euphrasia willkommii Freyn                               712     6.47  6.47
#>  4 Lotus corniculatus L. subsp. glacialis (Boiss.) Valdés   604     5.49  5.49
#>  5 Scorzoneroides                                           585     5.32  5.32
#>  6 Eleocharis quinqueflora (Hartmann) O.Schwarz             500     4.54  4.54
#>  7 Festuca iberica (Hack.) Patzke                           491     4.46  4.46
#>  8 Carex nevadensis Boiss. & Reut.                          482     4.38  4.38
#>  9 Gentiana boryi Boiss.                                    407     3.70  3.70
#> 10 Plantago nivalis Jord.                                   321     2.92  2.92
#> # ℹ 84 more rows
#> 
#> $genus
#> # A tibble: 52 × 4
#>    genus              n freq_all  freq
#>    <chr>          <int>    <dbl> <dbl>
#>  1 Carex           1824    16.6  16.6 
#>  2 Nardus          1029     9.35  9.35
#>  3 Scorzoneroides   990     9.00  9.00
#>  4 Gentiana         968     8.80  8.80
#>  5 Euphrasia        712     6.47  6.47
#>  6 Lotus            604     5.49  5.49
#>  7 Trifolium        586     5.33  5.33
#>  8 Festuca          525     4.77  4.77
#>  9 Eleocharis       500     4.54  4.54
#> 10 Agrostis         488     4.44  4.44
#> # ℹ 42 more rows
#>